access token included

+++ deletepost.php
@@ -7,31 +7,40 @@
 
 		$post_id = $_POST['post_id'];
 
+		$sql1 = "SELECT * FROM alumnus WHERE email=$_SESSION['email']";
+		$result1 = query($sql);
+		$row = fetch_array($result1);
+		$access_token = $row['access_token'];
+
 		$sql = "DELETE FROM posts WHERE id=$post_id";
-		$no=0;
+		$no=false;
 
 		$result = query($sql);
-
+        if($access_token==$_SESSION['access_token']){
 		 if(row_count($result)==1){
-		 $no=1;
+		 $no=true;
 		 $sql1 = "DELETE FROM threads WHERE post_id=$post_id";
 		 $result1 = query($sql1);
         
 	     }
 
-	     if($no==1){
+	     if($no){
 	     	$response['status']=209;
 	   $messages[]="post deleted succsesfully!";
 			
 
 		}
-		 if($n0==0){
+		 if($no){
 		 	$response['status']=408;
 			$messages[]="retry";
 			
 
 	
 		}
+	}else{
+		$response['status']=409;
+		$message[]="unauthorized access";
+	}
         
     
 	$response['messages'] = $messages;
